SmartAudio channel setting is recovered

+++ libraries/AP_VideoTX/AP_SmartAudio.h
@@ -236,7 +236,7 @@ private:
     // change the frequency
     void set_frequency(uint16_t frequency, bool isPitModeFreq);
     // change the channel that already includes its band
-    void set_channel(uint16_t chan);
+    void set_channel(uint8_t chan);
     // get the pitmode frequency
     void request_pit_mode_frequency();
     // set the power
